refactor(parser): add handler registry
