+++ src/parser/sshd_handlers.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "parser/handler_result.hpp"
+
+namespace loglens::parser_internal {
+
+HandlerResult handle_sshd_event(const Event& source);
+
+}  // namespace loglens::parser_internal
