+++ src/parser/su_handlers.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "parser/handler_result.hpp"
+
+namespace loglens::parser_internal {
+
+HandlerResult handle_su_event(const Event& source);
+
+}  // namespace loglens::parser_internal
